为第三方提供deal的api--india

+++ hasoffer.core/src/main/java/hasoffer/core/system/impl/AppServiceImpl.java
@@ -180,8 +180,8 @@ public class AppServiceImpl implements IAppService {
     }
 
     @Override
-    public PageableResult getDeals(Long page, Long pageSize) {
-        return dbm.queryPage(Q_APP_GETDEALS, page.intValue() <= 1 ? 1 : page.intValue(), pageSize.intValue(), Arrays.asList(new Date()));
+    public PageableResult getDeals(int page, int pageSize) {
+        return dbm.queryPage(Q_APP_GETDEALS, page <= 1 ? 1 : page, pageSize, Arrays.asList(new Date()));
 //        return dbm.queryPage(Q_APP_GETDEALS, page.intValue() <= 1 ? 1 : page.intValue(), pageSize.intValue(), Arrays.asList(new Date()));
     }
